+++ src/js/core/switcher.js
@@ -8,8 +8,7 @@ import {
     findIndex,
     getIndex,
     hasClass,
-    includes,
-    isNode,
+    isNumeric,
     isTag,
     matches,
     queryAll,
@@ -85,10 +84,10 @@ export default {
             this.$emit();
         },
 
-        toggles(toggles) {
+        toggles() {
             this.$emit();
             const active = this.index();
-            this.show(~active ? active : toggles[this.active] || toggles[0]);
+            this.show(~active ? active : this.next(this.active));
         },
     },
 
@@ -142,8 +141,7 @@ export default {
 
                 if (~i) {
                     e.preventDefault();
-                    const toggles = this.toggles.filter((el) => !matches(el, selDisabled));
-                    const next = toggles[getIndex(i, toggles, toggles.indexOf(current))];
+                    const next = this.toggles[this.next(i, this.toggles.indexOf(current))];
                     next.focus();
                     if (this.followFocus) {
                         this.show(next);
@@ -213,32 +211,43 @@ export default {
             return findIndex(this.children, (el) => hasClass(el, this.cls));
         },
 
-        show(item) {
+        next(item, prev = this.index()) {
+            if (isNumeric(item)) {
+                for (let i = 0; i < this.toggles.length; i++) {
+                    let index = getIndex(i + +item, this.toggles);
+                    if (!matches(this.toggles[index], selDisabled)) {
+                        return index;
+                    }
+                }
+            }
+
             const toggles = this.toggles.filter((el) => !matches(el, selDisabled));
-            const prev = this.index();
-            const next = getIndex(
-                !isNode(item) || includes(toggles, item) ? item : 0,
-                toggles,
-                getIndex(this.toggles[prev], toggles),
+            return getIndex(
+                toggles[getIndex(item, toggles, toggles.indexOf(this.toggles[prev]))],
+                this.toggles,
             );
-            const active = getIndex(toggles[next], this.toggles);
+        },
+
+        show(item) {
+            const prev = this.index();
+            const next = this.next(item);
 
             this.children.forEach((child, i) => {
-                toggleClass(child, this.cls, active === i);
+                toggleClass(child, this.cls, next === i);
                 attr(this.toggles[i], {
-                    'aria-selected': active === i,
-                    tabindex: active === i ? null : -1,
+                    'aria-selected': next === i,
+                    tabindex: next === i ? null : -1,
                 });
             });
 
             const animate = prev >= 0 && prev !== next;
             this.connects.forEach(async ({ children }) => {
                 const actives = toArray(children).filter(
-                    (child, i) => i !== active && hasClass(child, this.cls),
+                    (child, i) => i !== next && hasClass(child, this.cls),
                 );
 
                 if (await this.toggleElement(actives, false, animate)) {
-                    await this.toggleElement(children[active], true, animate);
+                    await this.toggleElement(children[next], true, animate);
                 }
             });
         },